Plugin working.
If there's any fix, misspelled words, method names or improvements,
please send an issue or a commit! <3

+++ new/src/main/java/com/lcastr0/prisonscoreboard/utils/ProgressBar.java
@@ -18,12 +18,12 @@ public class ProgressBar {
         this.character = character;
     }
 
-    public void setMax(long max){
-        this.max = max;
+    public void setMax(double max){
+        this.max = (long) max;
     }
 
-    public void setCurrent(long current){
-        this.current = current;
+    public void setCurrent(double current){
+        this.current = (long) current;
     }
 
     public String getProgressBar(){
